+++ src/main/java/com/cenxui/tea/app/service/admin/SettingController.java
@@ -2,5 +2,5 @@ package com.cenxui.tea.app.service.admin;
 
 import com.cenxui.tea.app.service.core.CoreController;
 
-public class AdminController extends CoreController {
+public class SettingController extends CoreController {
 }
