+++ src/main/java/com/cenxui/tea/app/service/admin/SignInController.java
@@ -1,0 +1,6 @@
+package com.cenxui.tea.app.service.admin;
+
+import com.cenxui.tea.app.service.core.CoreController;
+
+public class SignInController extends CoreController {
+}
