+++ src/main/java/com/cenxui/tea/app/service/customer/SettingController.java
@@ -1,0 +1,6 @@
+package com.cenxui.tea.app.service.customer;
+
+import com.cenxui.tea.app.service.core.CoreController;
+
+public class SettingController extends CoreController {
+}
