+++ src/main/java/com/cenxui/tea/app/service/user/UserController.java
@@ -1,27 +1,12 @@
 package com.cenxui.tea.app.service.user;
 
 import com.cenxui.tea.app.service.core.CoreController;
-import com.cenxui.tea.app.service.user.repository.UserRepository;
-import com.cenxui.tea.app.service.user.repository.UserRepositoryManager;
 import org.mindrot.jbcrypt.BCrypt;
 
 import java.util.List;
 
 public class UserController extends CoreController {
     private static final UserRepository manager = UserRepositoryManager.getManager();
-
-    public static boolean authenticateByUserName(String userName, String password) {
-        if (userName == null || userName.isEmpty()) return false;
-        if (password == null || password.isEmpty()) return false;
-
-        //use user name for query user
-        User user = manager.getUserByUserName(userName);
-        if (user == null) {
-            return false;
-        }
-        String hashedPassword = BCrypt.hashpw(password, user.getSalt());
-        return hashedPassword.equals(user.getHashedPassword());
-    }
 
     public static boolean authenticateByMail(String mail, String password) {
         if (mail == null || mail.isEmpty()) return false;
@@ -37,15 +22,6 @@ public class UserController extends CoreController {
 
     }
 
-    public static void setPasswordByUserName(String userName, String oldPassword, String newPassword) {
-        if (authenticateByUserName(userName, oldPassword)) {
-            String newSalt = BCrypt.gensalt();
-            String newHashedPassword = BCrypt.hashpw(newSalt, newPassword);
-            // Update the user salt and password by user
-            manager.setNewHashPasswordAndSaltByUser(userName, newSalt, newHashedPassword);
-        }
-    }
-
     public static void setPasswordByMail(String mail, String oldPassword, String newPassword) {
         if (authenticateByMail(mail, oldPassword)) {
             String newSalt = BCrypt.gensalt();
@@ -55,18 +31,25 @@ public class UserController extends CoreController {
         }
     }
 
-    public List<User> getUsers() {
-        //TODO
+    public static User getUserByMail(String mail) {
+        return manager.getUserByMail(mail);
+    }
+
+
+    public static List<User> getUsers() {
+        //TODO mia
         return null;
     }
 
-    public static boolean addUser() {
-        //TODO
+    public static boolean addUser(User user) {
+        //TODO mia
         return false;
     }
 
-    public static boolean removeUser() {
-        //TODO
+    public static boolean removeUser(User user) {
+        //TODO mia
         return false;
     }
+
+
  }
