+++ src/test/java/com/cenxui/tea/app/service/user/UserControllerTest.java
@@ -1,6 +1,5 @@
 package com.cenxui.tea.app.service.user;
 
-import com.cenxui.tea.app.service.user.repository.MockUserRepository;
 import com.cenxui.tea.app.util.ControllerManagerUtil;
 import org.junit.Before;
 import org.junit.Test;
@@ -12,11 +11,6 @@ public class UserControllerTest {
         ControllerManagerUtil.mockControllerManager(UserController.class, "manager", new MockUserRepository());
     }
 
-
-    @Test
-    public void authenticateByUserName() throws Exception {
-        System.out.println(UserController.authenticateByUserName("userName", "password"));
-    }
 
     @Test
     public void authenticateByMail() throws Exception {
